fix warning

+++ container.cpp
@@ -13,7 +13,7 @@ Container::Container(vector<Widget*> children)
 }
 
 void Container::handle_children_events(event ev) {
-    if(_children.size() < _focus || _children.size() == 0)
+    if((int)_children.size() < _focus || _children.size() == 0)
         _focus = -1;
 
     int old_focus = _focus;
